Non-reused geometry can now be received from the server in a GPU-ready format. This saves us from doing transformations on the CPU.
This is a settings which is disabled by default, you can enable it in
dev.js by uncommenting. The feature seems to work in both layers,
including selection/hide/color change. It does not yet work with reuse.

- 2 new geometryTypes in GeometryLoader
- Changed calculateBytesUsed signature
- geometryLoader has a reference to unquantizationMatrix, which is a lot
of cases is not specific to the octree-node
- RenderLayer.addCompleteBuffer added to allow for inserting a GPU
buffer directly (without am intermediate BufferSet). Also overridden in
both subclasses.
- Added a few more methods to vertexquantization (needs to be cleanup
up)