Several refactorings to make viewer more dynamic/adjustable
BimServerViewer doesn't require the BimServerApi to be given in the
constructor anymore. This way it is possible to load models from
different BIMservers in the same view. Consequently the loadModel and
other load methods now require ah api argument.

Introduced AbstractViewer class as a supertype of BimServerViewer. The
idea is that other viewers can also subclass AbstractViewer (and that
AbstractViewer becomes independant of BIMserver and API).