Changed containedInstances size from 1024 to 256. According to https://webglstats.com/webgl2/parameter/MAX_VERTEX_UNIFORM_VECTORS this amount of uniforms seems to be supported by all platforms.
In the future this needs to become variable, so we can use more on
higher-end GPUs, and possibly even less (or disable instancing) on
low-spec machines. Also, for performance reasons, making this variable
based on the actual amount of instancing that is applied would be even
better.

This fixes https://github.com/opensourceBIM/BIMsurfer/issues/64 at least
on my mac.